Got select working properly
